+++ e2e-tests/production-runtime/cypress/integration/client-only-paths.js
@@ -37,6 +37,11 @@ describe(`Client only paths`, () => {
       it(label, () => {
         cy.visit(path).waitForRouteChange()
         cy.getTestElement(`dom-marker`).contains(marker)
+
+        // Skipping these at the moment because behaviour is inconsistent
+        // https://github.com/gatsbyjs/gatsby/issues/16533
+        if (marker !== `index` && marker !== `static-sibling`)
+          cy.url().should(`eq`, Cypress.config().baseUrl + path)
       })
     })
   })
@@ -49,6 +54,7 @@ describe(`Client only paths`, () => {
       it(label, () => {
         cy.navigateAndWaitForRouteChange(path)
         cy.getTestElement(`dom-marker`).contains(marker)
+        cy.url().should(`eq`, Cypress.config().baseUrl + path)
       })
     })
   })