+++ packages/gatsby-cli/src/reporter/index.js
@@ -38,6 +38,13 @@ const reporter: Reporter = {
     if (isNoColor) {
       errorFormatter.withoutColors()
     }
+
+    // disables colors in popular terminal output coloring packages
+    //  - chalk: see https://www.npmjs.com/package/chalk#chalksupportscolor
+    //  - ansi-colors: see https://github.com/doowb/ansi-colors/blob/8024126c7115a0efb25a9a0e87bc5e29fd66831f/index.js#L5-L7
+    if (isNoColor) {
+      process.env.FORCE_COLOR = `0`
+    }
   },
   /**
    * Log arguments and exit process with status 1.
@@ -227,8 +234,8 @@ const reporter: Reporter = {
 }
 
 console.log = (...args) => reporter.log(util.format(...args))
-console.warn = (...args) => reporter.log(util.format(...args))
-console.info = (...args) => reporter.log(util.format(...args))
-console.error = (...args) => reporter.log(util.format(...args))
+console.warn = (...args) => reporter.warn(util.format(...args))
+console.info = (...args) => reporter.info(util.format(...args))
+console.error = (...args) => reporter.error(util.format(...args))
 
 module.exports = reporter