+++ packages/gatsby-plugin-mdx/component-with-mdx-scope.js
@@ -2,8 +2,6 @@
 // turn function into a no-op passthrough
 // this is so that people don't really have to change their code
 module.exports = function componentWithMDXScope(componentPath) {
-  console.warn(
-    `componentWithMDXScope is deprecated. You don't need it anymore`
-  );
-  return componentPath;
-};
+  console.warn(`componentWithMDXScope is deprecated. You don't need it anymore`)
+  return componentPath
+}