+++ packages/gatsby-remark-images/src/constants.js
@@ -4,8 +4,11 @@ exports.DEFAULT_OPTIONS = {
   backgroundColor: `white`,
   linkImagesToOriginal: true,
   showCaptions: false,
+  markdownCaptions: false,
   withWebp: false,
   tracedSVG: false,
+  loading: `lazy`,
+  disableBgImageOnAlpha: false,
 }
 
 exports.imageClass = `gatsby-resp-image-image`