+++ packages/gatsby-remark-prismjs/src/highlight-code.js
@@ -41,6 +41,7 @@ module.exports = (
   const grammar = Prism.languages[language]
   const highlighted = Prism.highlight(code, grammar, language)
   const codeSplits = handleDirectives(highlighted, lineNumbersHighlight)
+
   let finalCode = ``
   const lastIdx = codeSplits.length - 1 // Don't add back the new line character after highlighted lines
   // as they need to be display: block and full-width.