+++ packages/gatsby/cache-dir/loader.js
@@ -1,7 +1,6 @@
-import "core-js/modules/es7.promise.finally"
 import prefetchHelper from "./prefetch"
 import emitter from "./emitter"
-import { setMatchPaths, findMatchPath, cleanPath } from "./find-path"
+import { setMatchPaths, findPath, findMatchPath } from "./find-path"
 
 const preferDefault = m => (m && m.default) || m
 
@@ -20,7 +19,6 @@ const doFetch = (url, method = `GET`) =>
   new Promise((resolve, reject) => {
     const req = new XMLHttpRequest()
     req.open(method, url, true)
-    req.withCredentials = true
     req.onreadystatechange = () => {
       if (req.readyState == 4) {
         resolve(req)
@@ -88,7 +86,10 @@ const loadPageDataJson = loadObj => {
 }
 
 const doesConnectionSupportPrefetch = () => {
-  if (`connection` in navigator) {
+  if (
+    `connection` in navigator &&
+    typeof navigator.connection !== `undefined`
+  ) {
     if ((navigator.connection.effectiveType || ``).includes(`2g`)) {
       return false
     }
@@ -144,7 +145,7 @@ export class BaseLoader {
   }
 
   loadPageDataJson(rawPath) {
-    const pagePath = cleanPath(rawPath)
+    const pagePath = findPath(rawPath)
     if (this.pageDataDb.has(pagePath)) {
       return Promise.resolve(this.pageDataDb.get(pagePath))
     }
@@ -162,7 +163,7 @@ export class BaseLoader {
 
   // TODO check all uses of this and whether they use undefined for page resources not exist
   loadPage(rawPath) {
-    const pagePath = cleanPath(rawPath)
+    const pagePath = findPath(rawPath)
     if (this.pageDb.has(pagePath)) {
       const page = this.pageDb.get(pagePath)
       return Promise.resolve(page.payload)
@@ -173,18 +174,6 @@ export class BaseLoader {
 
     const inFlight = this.loadPageDataJson(pagePath)
       .then(result => {
-        if (result.notFound) {
-          // if request was a 404, we should fallback to findMatchPath.
-          let foundMatchPatch = findMatchPath(pagePath)
-          if (foundMatchPatch && foundMatchPatch !== pagePath) {
-            return this.loadPage(foundMatchPatch).then(pageResources => {
-              this.pageDb.set(pagePath, this.pageDb.get(foundMatchPatch))
-
-              return pageResources
-            })
-          }
-        }
-
         if (result.status === `error`) {
           return {
             status: `error`,
@@ -221,8 +210,14 @@ export class BaseLoader {
           return pageResources
         })
       })
-      .finally(() => {
+      // prefer duplication with then + catch over .finally to prevent problems in ie11 + firefox
+      .then(response => {
         this.inFlightDb.delete(pagePath)
+        return response
+      })
+      .catch(err => {
+        this.inFlightDb.delete(pagePath)
+        throw err
       })
 
     this.inFlightDb.set(pagePath, inFlight)
@@ -231,7 +226,7 @@ export class BaseLoader {
 
   // returns undefined if loading page ran into errors
   loadPageSync(rawPath) {
-    const pagePath = cleanPath(rawPath)
+    const pagePath = findPath(rawPath)
     if (this.pageDb.has(pagePath)) {
       return this.pageDb.get(pagePath).payload
     }
@@ -269,18 +264,10 @@ export class BaseLoader {
       return false
     }
 
-    const realPath = cleanPath(pagePath)
+    const realPath = findPath(pagePath)
     // Todo make doPrefetch logic cacheable
     // eslint-disable-next-line consistent-return
-    this.doPrefetch(realPath).then(pageData => {
-      if (!pageData) {
-        const matchPath = findMatchPath(realPath)
-
-        if (matchPath && matchPath !== realPath) {
-          return this.prefetch(matchPath)
-        }
-      }
-
+    this.doPrefetch(realPath).then(() => {
       if (!this.prefetchCompleted.has(pagePath)) {
         this.apiRunner(`onPostPrefetchPathname`, { pathname: pagePath })
         this.prefetchCompleted.add(pagePath)
@@ -299,7 +286,7 @@ export class BaseLoader {
   }
 
   getResourceURLsForPathname(rawPath) {
-    const pagePath = cleanPath(rawPath)
+    const pagePath = findPath(rawPath)
     const page = this.pageDataDb.get(pagePath)
     if (page) {
       const pageResources = toPageResources(page.payload)
@@ -314,7 +301,7 @@ export class BaseLoader {
   }
 
   isPageNotFound(rawPath) {
-    const pagePath = cleanPath(rawPath)
+    const pagePath = findPath(rawPath)
     const page = this.pageDb.get(pagePath)
     return page && page.notFound === true
   }