+++ packages/gatsby/src/redux/reducers/schema-customization.js
@@ -1,7 +1,9 @@
 module.exports = (
   state = {
     composer: null,
+    context: {},
     fieldExtensions: {},
+    printConfig: null,
     thirdPartySchemas: [],
     types: [],
   },
@@ -48,10 +50,31 @@ module.exports = (
         fieldExtensions: { ...state.fieldExtensions, [name]: extension },
       }
     }
+    case `PRINT_SCHEMA_REQUESTED`: {
+      const { path, include, exclude, withFieldTypes } = action.payload
+      return {
+        ...state,
+        printConfig: {
+          path,
+          include,
+          exclude,
+          withFieldTypes,
+        },
+      }
+    }
+    case `CREATE_RESOLVER_CONTEXT`: {
+      const context = action.payload
+      return {
+        ...state,
+        context: { ...state.context, ...context },
+      }
+    }
     case `DELETE_CACHE`:
       return {
         composer: null,
+        context: {},
         fieldExtensions: {},
+        printConfig: null,
         thirdPartySchemas: [],
         types: [],
       }