+++ packages/gatsby/src/schema/__tests__/queries.js
@@ -1,5 +1,6 @@
 const { graphql } = require(`graphql`)
 const { store } = require(`../../redux`)
+const { actions } = require(`../../redux/actions`)
 const { build } = require(`..`)
 const withResolverContext = require(`../context`)
 require(`../../db/__tests__/fixtures/ensure-loki`)()
@@ -9,11 +10,36 @@ const apiRunnerNode = require(`../../utils/api-runner-node`)
 
 const nodes = require(`./fixtures/queries`)
 
+jest.mock(`gatsby-cli/lib/reporter`, () => {
+  return {
+    log: jest.fn(),
+    info: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+    activityTimer: () => {
+      return {
+        start: jest.fn(),
+        setStatus: jest.fn(),
+        end: jest.fn(),
+      }
+    },
+  }
+})
+
 describe(`Query schema`, () => {
   let schema
+  let schemaComposer
 
   const runQuery = query =>
-    graphql(schema, query, undefined, withResolverContext({}, schema))
+    graphql(
+      schema,
+      query,
+      undefined,
+      withResolverContext({
+        schema,
+        schemaComposer,
+      })
+    )
 
   beforeAll(async () => {
     apiRunnerNode.mockImplementation(async (api, ...args) => {
@@ -118,7 +144,7 @@ describe(`Query schema`, () => {
 
     store.dispatch({ type: `DELETE_CACHE` })
     nodes.forEach(node =>
-      store.dispatch({ type: `CREATE_NODE`, payload: node })
+      actions.createNode(node, { name: `test` })(store.dispatch)
     )
 
     const typeDefs = [
@@ -141,6 +167,7 @@ describe(`Query schema`, () => {
 
     await build({})
     schema = store.getState().schema
+    schemaComposer = store.getState().schemaCustomization.composer
   })
 
   describe(`on children fields`, () => {