+++ themes/gatsby-theme-blog/src/gatsby-plugin-theme-ui/styles.js
@@ -7,7 +7,7 @@ export default {
     fontFamily: `monospace`,
     tabSize: 4,
     hyphens: `none`,
-    marginBottom: 0,
+    marginBottom: 2,
     color: `white`,
     bg: `prism.background`,
     overflow: `auto`,