+++ themes/gatsby-theme-notes/src/components/breadcrumbs.js
@@ -7,7 +7,7 @@ import BreadcrumbDivider from "./breadcrumb-divider"
 import BreadcrumbHome from "./breadcrumb-home"
 
 export default ({ links }) => {
-  const { basePath = `/`, homeText, breadcrumbSeparator } = useOptions()
+  const { homeText, breadcrumbSeparator } = useOptions()
 
   return (
     <nav
@@ -20,11 +20,6 @@ export default ({ links }) => {
       })}
     >
       <BreadcrumbHome text={homeText} />
-      <BreadcrumbDivider text={breadcrumbSeparator} />
-
-      <Styled.a as={Link} to={basePath}>
-        {basePath.replace(/^\//, ``)}
-      </Styled.a>
       {links.map(link => (
         <>
           <BreadcrumbDivider text={breadcrumbSeparator} />