+++ www/src/components/api-reference/doc-block.js
@@ -47,17 +47,15 @@ const Deprecated = ({ definition }) => {
   return null
 }
 
-const APILink = ({ definition, githubPath }) => {
-  if (definition.codeLocation && githubPath) {
+const APILink = ({ definition, relativeFilePath }) => {
+  if (definition.codeLocation && relativeFilePath) {
     return (
       <a
         sx={{
           ...linkStyles,
           display: `inline-flex !important`,
         }}
-        href={`${githubPath}#L${definition.codeLocation.start.line}-L${
-          definition.codeLocation.end.line
-        }`}
+        href={`https://github.com/gatsbyjs/gatsby/blob/${process.env.COMMIT_SHA}/packages/${relativeFilePath}#L${definition.codeLocation.start.line}-L${definition.codeLocation.end.line}`}
         aria-label="View source on GitHub"
       >
         <GithubIcon focusable="false" sx={{ mr: 2 }} />
@@ -68,7 +66,7 @@ const APILink = ({ definition, githubPath }) => {
 
   if (
     definition.codeLocation &&
-    !githubPath &&
+    !relativeFilePath &&
     !Array.isArray(definition.codeLocation)
   ) {
     return (
@@ -77,11 +75,7 @@ const APILink = ({ definition, githubPath }) => {
           ...linkStyles,
           display: `inline-flex !important`,
         }}
-        href={`https://github.com/gatsbyjs/gatsby/blob/${
-          process.env.COMMIT_SHA
-        }/packages/${definition.codeLocation.file}#L${
-          definition.codeLocation.start.line
-        }-L${definition.codeLocation.end.line}`}
+        href={`https://github.com/gatsbyjs/gatsby/blob/${process.env.COMMIT_SHA}/packages/${definition.codeLocation.file}#L${definition.codeLocation.start.line}-L${definition.codeLocation.end.line}`}
         aria-label="View source on GitHub"
       >
         <GithubIcon focusable="false" sx={{ mr: 2 }} />
@@ -92,7 +86,7 @@ const APILink = ({ definition, githubPath }) => {
 
   if (
     definition.codeLocation &&
-    !githubPath &&
+    !relativeFilePath &&
     Array.isArray(definition.codeLocation)
   ) {
     return (
@@ -112,9 +106,7 @@ const APILink = ({ definition, githubPath }) => {
           {definition.codeLocation.map((loc, index) => (
             <LinkBox
               key={`${loc.file}${loc.start.line}-${loc.end.line}`}
-              href={`https://github.com/gatsbyjs/gatsby/blob/${
-                process.env.COMMIT_SHA
-              }/packages/${loc.file}#L${loc.start.line}-L${loc.end.line}`}
+              href={`https://github.com/gatsbyjs/gatsby/blob/${process.env.COMMIT_SHA}/packages/${loc.file}#L${loc.start.line}-L${loc.end.line}`}
               aria-label={`View source #${index + 1} on GitHub`}
             >
               {index + 1}
@@ -176,7 +168,7 @@ const Description = ({ definition }) => {
 
 const DocBlock = ({
   definition,
-  githubPath = null,
+  relativeFilePath = null,
   level = 0,
   linkableTitle = false,
   title = null,
@@ -230,7 +222,10 @@ const DocBlock = ({
           )}
         </div>
         {level === 0 && (
-          <APILink githubPath={githubPath} definition={definition} />
+          <APILink
+            relativeFilePath={relativeFilePath}
+            definition={definition}
+          />
         )}
         {definition.optional && <Optional />}
       </Header>