+++ www/src/components/docs-table-of-contents.js
@@ -4,7 +4,18 @@ import React from "react"
 import { Link } from "gatsby"
 import { mediaQueries } from "../gatsby-plugin-theme-ui"
 
-function createItems(items, location) {
+function isUnderDepthLimit(depth, maxDepth) {
+  if (maxDepth === null) {
+    // if no maxDepth is passed in, continue to render more items
+    return true
+  } else {
+    return depth < maxDepth
+  }
+}
+
+// depth and maxDepth are used to figure out how many bullets deep to render in the ToC sidebar, if no
+// max depth is set via the tableOfContentsDepth field in the frontmatter, all headings will be rendered
+function createItems(items, location, depth, maxDepth) {
   return (
     items &&
     items.map(item => (
@@ -25,12 +36,24 @@ function createItems(items, location) {
               },
             },
           }}
+          getProps={({ href, location }) =>
+            location && location.href && location.href.includes(href)
+              ? {
+                  style: {
+                    color: colors.link.color,
+                    borderBottom: `1px solid ${colors.link.hoverBorder}`,
+                  },
+                }
+              : null
+          }
           to={location.pathname + item.url}
         >
           {item.title}
         </Link>
-        {item.items && (
-          <ul sx={{ ml: 6 }}>{createItems(item.items, location)}</ul>
+        {item.items && isUnderDepthLimit(depth, maxDepth) && (
+          <ul sx={{ ml: 6 }}>
+            {createItems(item.items, location, depth + 1, maxDepth)}
+          </ul>
         )}
       </li>
     ))
@@ -59,7 +82,12 @@ function TableOfContents({ page, location }) {
           },
         }}
       >
-        {createItems(page.tableOfContents.items, location)}
+        {createItems(
+          page.tableOfContents.items,
+          location,
+          1,
+          page.frontmatter.tableOfContentsDepth
+        )}
       </ul>
     </nav>
   ) : null