+++ packages/gatsby-core-utils/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.10...gatsby-core-utils@1.0.12) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.10...gatsby-core-utils@1.0.11) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.9...gatsby-core-utils@1.0.10) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.8...gatsby-core-utils@1.0.9) (2019-09-18)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.7...gatsby-core-utils@1.0.8) (2019-09-09)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.6...gatsby-core-utils@1.0.7) (2019-09-01)
+
+### Bug Fixes
+
+- update minor updates in packages except react, babel and eslint ([#17254](https://github.com/gatsbyjs/gatsby/issues/17254)) ([252d867](https://github.com/gatsbyjs/gatsby/commit/252d867))
+
+## [1.0.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.5...gatsby-core-utils@1.0.6) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.4...gatsby-core-utils@1.0.5) (2019-08-20)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.3...gatsby-core-utils@1.0.4) (2019-08-06)
+
+### Bug Fixes
+
+- **www:** replace \n into newlines for guess ([#16382](https://github.com/gatsbyjs/gatsby/issues/16382)) ([44dbafb](https://github.com/gatsbyjs/gatsby/commit/44dbafb))
+
 ## [1.0.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.2...gatsby-core-utils@1.0.3) (2019-07-12)
 
 **Note:** Version bump only for package gatsby-core-utils